feat: Model 側の Interface を更新

+++ Assets/DQEmulation/Scripts/00_Domain/Model/Interface/ICreature.cs
@@ -7,9 +7,9 @@ namespace Denik.DQEmulation.Model
     {
         string Name { get; }
         Sprite Figure { get; }
-        IReadOnlyReactiveProperty<int> HitPoint { get; }
+        IReadOnlyReactiveProperty<float> HitPoint { get; }
         // HP最大値が上昇する場合，ReactiveProperty 化するのがよい
-        int MaxHitPoint { get; }
+        float MaxHitPoint { get; }
         // 今回はMPを省ているが，HPと同じ要領で定義する
     }
 }
